Fix remaining review issues: JSON unicode uppercase, APR_CREATE file creation, LLError ignoreCase, thread pool, TextureAnim UByte comparison, LLCircuit seconds comparison
Agent-Logs-Url: https://github.com/Kaleaon/Firestorm-kotlin/sessions/b29c64e5-0c74-4d72-87c8-f9ea01f1c764

+++ src/main/kotlin/com/firestorm/llprimitive/TextureAnim.kt
@@ -157,8 +157,8 @@ data class TextureAnim(
         rate   = buf.float
         // Clamp tile sizes: must be >= 1 unless TAM_SMOOTH is set
         if (!isSmooth()) {
-            if (sizeX == 0u.toUByte()) sizeX = 1u
-            if (sizeY == 0u.toUByte()) sizeY = 1u
+            if (sizeX.toUInt() == 0u) sizeX = 1u
+            if (sizeY.toUInt() == 0u) sizeY = 1u
         }
     }
 }
